Move layouts to components folder
