+++ src/components/layout/DefaultLayout.tsx
@@ -1,13 +1,14 @@
 import { FunctionComponent } from 'react';
-import PublicHeader from '../components/PublicHeader';
+
+import PublicHeader from '../PublicHeader';
 
 import { Content, Flex } from '@adobe/react-spectrum';
 
 const DefaultLayout : FunctionComponent = ({ children }) => (
     <Flex
-        direction='column'
-        gap='size-100'
-        margin='size-200'>
+        direction="column"
+        gap="size-100"
+        margin="size-200">
         <PublicHeader user={ null }/>
         <Content>
             { children }
