+++ src/components/layout/OrgLayout.tsx
@@ -4,7 +4,7 @@ import { useRouter } from 'next/router';
 import { Item, Tabs } from '@react-spectrum/tabs';
 
 import DefaultLayout from './DefaultLayout';
-import getOrg from '../fetching/getOrg';
+import getOrg from '../../fetching/getOrg';
 import OrgHeader from './OrgHeader';
 
 interface OrgLayoutProps {
@@ -26,16 +26,16 @@ const OrgLayout = ({ children, orgId } : OrgLayoutProps) : JSX.Element => {
         <DefaultLayout>
             <OrgHeader org={ orgQuery.data }/>
             <Tabs
-                aria-label='Organization submenu'
+                aria-label="Organization submenu"
                 selectedKey={ currentTab }
                 onSelectionChange={ onSelectTab }>
-                <Item title='Coming up' key='events'>
+                <Item title="Coming up" key="events">
                     <Content>{ children }</Content>
                 </Item>
-                <Item title='Contact' key='contact'>
+                <Item title="Contact" key="contact">
                     <Content>{ children }</Content>
                 </Item>
-                <Item title='Some Custom Page' key='custom'>
+                <Item title="Some Custom Page" key="custom">
                     <Content>{ children }</Content>
                 </Item>
             </Tabs>
