+++ src/pages/o/[orgId]/contact.tsx
@@ -3,7 +3,7 @@ import { GetServerSideProps } from 'next';
 import { QueryClient } from 'react-query';
 
 import getOrg from '../../../fetching/getOrg';
-import OrgLayout from '../../../layout/OrgLayout';
+import OrgLayout from '../../../components/layout/OrgLayout';
 
 export const getServerSideProps : GetServerSideProps = async (context) => {
     const queryClient = new QueryClient();
